COMP: Fix -Wsign-promo in Widgets & DICOMWidgets
This commit fixes warnings like the following:

  /path/to/CTK/Libs/Widgets/Testing/Cpp/ctkCheckableHeaderViewTest1.cpp:76:73: warning: passing ‘Qt::CheckState’ chooses ‘int’ over ‘uint’ {aka ‘unsigned int’} [-Wsign-promo]
     76 |   model.setHeaderData(0, Qt::Horizontal, Qt::Checked, Qt::CheckStateRole);
        |

+++ Libs/DICOM/Widgets/Testing/Cpp/ctkDICOMModelTest2.cpp
@@ -87,7 +87,7 @@ int ctkDICOMModelTest2( int argc, char * argv [] )
     headerView->checkableModelHelper()->setPropagateDepth(-1);
     headerView->checkableModelHelper()->setForceCheckability(true);
     viewer.setHeader(headerView);
-    model.setHeaderData(0, Qt::Horizontal, Qt::Checked, Qt::CheckStateRole);
+    model.setHeaderData(0, Qt::Horizontal, static_cast<int>(Qt::Checked), Qt::CheckStateRole);
     qDebug() << "new: " << headerView->isHidden();
     topLevel.show();
     if (argc <= 3 || QString(argv[3]) != "-I")
