+++ Libs/Widgets/Testing/Cpp/ctkCheckableHeaderViewEventTranslatorPlayerTest1.cpp
@@ -103,7 +103,7 @@ int ctkCheckableHeaderViewEventTranslatorPlayerTest1(int argc, char * argv [] )
   table.setModel(&model);
 
   // Header is checked by default
-  model.setHeaderData(0, Qt::Horizontal, Qt::Checked, Qt::CheckStateRole);
+  model.setHeaderData(0, Qt::Horizontal, static_cast<int>(Qt::Checked), Qt::CheckStateRole);
 
   QHeaderView* previousHeaderView = table.horizontalHeader();
 #if QT_VERSION < QT_VERSION_CHECK(5,0,0)
@@ -128,7 +128,7 @@ int ctkCheckableHeaderViewEventTranslatorPlayerTest1(int argc, char * argv [] )
   table.setHorizontalHeader(headerView);
 
   // Header is checked by default
-  model.setHeaderData(1, Qt::Horizontal, Qt::Checked, Qt::CheckStateRole);
+  model.setHeaderData(1, Qt::Horizontal, static_cast<int>(Qt::Checked), Qt::CheckStateRole);
 
   QSignalSpy spy1(headerView, SIGNAL(sectionPressed(int)));
